+++ src/themes/a4-poster.ts
@@ -1,0 +1,127 @@
+import { Theme } from "../types";
+
+export const a4PosterTheme: Theme = {
+  name: "a4-poster",
+  wrapperStyle: {
+    width: "794px",
+    height: "1123px",
+    backgroundColor: "#f5f7fa",
+    fontFamily: '"Inter", sans-serif',
+    color: "#000",
+    textAlign: "left",
+    padding: "80px",
+    boxSizing: "border-box",
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+  },
+  elementStyles: {
+    h1: {
+      fontFamily: '"Playfair Display", serif',
+      fontSize: "52px",
+      fontWeight: "600",
+      margin: "0 0 8px 0",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    h2: {
+      fontFamily: '"Playfair Display", serif',
+      fontSize: "52px",
+      fontStyle: "italic",
+      fontWeight: "600",
+      margin: "0 0 40px 0",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    h3: {
+      fontFamily: '"Playfair Display", serif',
+      fontSize: "36px",
+      fontWeight: "600",
+      margin: "1em 0 0.5em",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    h4: {
+      fontFamily: '"Playfair Display", serif',
+      fontSize: "32px",
+      fontWeight: "600",
+      margin: "1em 0 0.5em",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    h5: {
+      fontFamily: '"Playfair Display", serif',
+      fontSize: "28px",
+      fontWeight: "600",
+      margin: "1em 0 0.5em",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    h6: {
+      fontFamily: '"Playfair Display", serif',
+      fontSize: "24px",
+      fontWeight: "600",
+      margin: "1em 0 0.5em",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    p: {
+      fontSize: "20px",
+      lineHeight: "1.75",
+      marginBottom: "24px",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    em: { fontStyle: "italic" },
+    strong: { fontWeight: "bold" },
+    u: { textDecoration: "underline" },
+    s: { textDecoration: "line-through" },
+    blockquote: {
+      marginLeft: "2rem",
+      paddingLeft: "1.5rem",
+      borderLeft: "3px solid #000",
+      fontStyle: "italic",
+      color: "#333",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    code: {
+      fontFamily: "monospace",
+      backgroundColor: "#edf2f7",
+      padding: "0.2em 0.4em",
+      borderRadius: "4px",
+      fontSize: "0.95em",
+    },
+    pre: {
+      backgroundColor: "#edf2f7",
+      padding: "1.5em",
+      borderRadius: "6px",
+      overflowX: "auto",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    ul: {
+      marginLeft: "1.5rem",
+      paddingLeft: "1rem",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    ol: {
+      marginLeft: "1.5rem",
+      paddingLeft: "1rem",
+      width: "100%",
+      maxWidth: "650px",
+    },
+    li: { marginBottom: "0.5em", fontSize: "20px" },
+    a: { color: "#2b6cb0", textDecoration: "underline" },
+    div: {
+      fontFamily: '"Playfair Display", serif',
+      fontSize: "20px",
+      fontWeight: "bold",
+      letterSpacing: "2px",
+      marginTop: "40px",
+      width: "100%",
+      maxWidth: "650px",
+    },
+  },
+};
