+++ src/themes/index.ts
@@ -1,0 +1,10 @@
+import { Theme } from "../types";
+import { bookExcerptTheme } from "./book-excerpt";
+import { a4PosterTheme } from "./a4-poster";
+import { highlightedBookTheme } from "./highlighted-book";
+
+export const themes: Theme[] = [
+  bookExcerptTheme,
+  a4PosterTheme,
+  highlightedBookTheme,
+];
